requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
twilio
psycopg2-binary
